requirements.txt
python-cinderclient
